# ⚡ Optimize `isAssignable` for literal types