The vast majority (if not all) calls to `isAssignable` compare two literal types. This PR introduces a fast path for those cases, assuming that two literal types of the same kind are always equal to each other when their values match.

A few benchmarks:
* 9.39ms -> 2.36ms
* 170.11ms -> 117.26ms
* 320.82ms -> 274.17ms

That last one in particular was
```
data merge entity @s {Air:1b,state:"rolling",Passengers:[{Air:0}]}
```

Additionally this PR removes some of the attribute warnings, bringing the total time of that last example all the way down to 8.08ms!