@misode 2024-06-15
These are too good not to merge, sorry. If this literal isAssignable check turns out to be wrong, blame me later.