@NeunEinser 2024-06-16
> Additionally this PR removes some of the attribute warnings, bringing the total time of that last example all the way down to 8.08ms!

Would this be smaller when all attributes are implemented? I think I prefer having warnings here.