@misode 2024-06-16
> Would this be smaller when all attributes are implemented? I think I prefer having warnings here.

There were a lot of duplicate warnings, all in the hot simplify path. If we want warnings for unimplemented attributes, I think it would be better to do this during the mcdoc binding process, and also de-duplicate them.